+++ design-system/avatar-group/src/components/AvatarGroup.tsx
@@ -89,17 +89,15 @@ export interface AvatarGroupProps {
    */
   boundariesElement?: 'viewport' | 'window' | 'scrollParent';
 
-  // Add these in the next major release
-  // boundary?: 'clippingParents' | HTMLElement;
-  // rootBoundary?: 'viewport' | 'document';
-
   /**
    * A `testId` prop is provided for specified elements,
    * which is a unique string that appears as a data attribute `data-testid` in the rendered code,
    * serving as a hook for automated tests.
+   */
+  //
 
+  /**
    * Will set these elements when defined:
-
    * - Container element - `{testId}--avatar-group`
    * - Avatar items - `{testId}--avatar-{index}`
    * - Overflow menu button - `{testId}--overflow-menu--trigger`