+++ design-system/avatar/examples/constellation/avatar-circle.tsx
@@ -8,6 +8,7 @@ export default function AvatarCircleExample() {
       appearance="circle"
       src="https://pbs.twimg.com/profile_images/803832195970433027/aaoG6PJI_400x400.jpg"
       size="large"
+      name="John Doe"
     />
   );
 }