+++ editor/editor-core/example-helpers/DevTools.tsx
@@ -1,22 +1,36 @@
 import React from 'react';
 import { WithEditorActions } from '../src';
 
-export function DevTools() {
-  if (process.env.NODE_ENV !== 'test') {
-    const workerFile = require('worker-plugin/loader!prosemirror-dev-tools/json-diff.worker');
-    // eslint-disable-next-line react-hooks/rules-of-hooks
-    const diffWorker = React.useMemo(
-      () =>
-        typeof Worker === 'undefined' ? undefined : new Worker(workerFile),
-      [workerFile],
-    );
+const useDiffWorker = () => {
+  let workerFile: string | null;
 
+  try {
+    workerFile = require('worker-plugin/loader!prosemirror-dev-tools/json-diff.worker');
+  } catch (err) {
+    workerFile = null;
+  }
+
+  return React.useMemo(
+    () =>
+      typeof Worker === 'undefined' || !workerFile
+        ? undefined
+        : new Worker(workerFile),
+    [workerFile],
+  );
+};
+
+export function DevTools() {
+  const diffWorker = useDiffWorker();
+
+  if (process.env.NODE_ENV !== 'test') {
     return (
       <WithEditorActions
         render={actions => {
           const editorView = actions._privateGetEditorView();
           if (editorView) {
-            import('prosemirror-dev-tools').then(({ applyDevTools }) =>
+            import(
+              /* webpackChunkName: "@atlaskit-internal_prosemirror-dev-tools" */ 'prosemirror-dev-tools'
+            ).then(({ applyDevTools }) =>
               applyDevTools(editorView, {
                 diffWorker,
               }),