+++ elements/user-picker/src/__tests__/unit/components/AvatarItemOptionSpec.tsx
@@ -1,18 +1,31 @@
 import { mount } from 'enzyme';
-import React from 'react';
+import React, { ReactNode } from 'react';
+import { LozengeProps } from '../../../types';
 import { AvatarItemOption } from '../../../components/AvatarItemOption';
 
-type Lozenge = {
-  text: string;
-  appearance: 'inprogress';
-};
+jest.mock('@atlaskit/tooltip', () => ({
+  ...jest.requireActual<any>('@atlaskit/tooltip'),
+  __esModule: true,
+  default: ({
+    children,
+    content,
+  }: {
+    children: ReactNode;
+    content: string;
+  }) => (
+    <div>
+      <div>{children}</div>
+      <div>{content}</div>
+    </div>
+  ),
+}));
 
 describe('AvatarItemOption', () => {
   describe('should render AvatarItem with', () => {
     const primaryText = 'PrimaryText';
     const secondaryText = 'SecondaryText';
     const sourcesInfoTooltip = 'Sources Info Tooltip';
-    const lozenge: Lozenge = {
+    const lozenge: LozengeProps = {
       text: 'in progress',
       appearance: 'inprogress',
     };
@@ -78,6 +91,30 @@ describe('AvatarItemOption', () => {
 
         expect(component.text()).not.toContain(sourcesInfoTooltip);
         expect(component.text()).toContain(lozenge.text);
+      },
+    );
+
+    it(
+      'with lozenge and lozenge tooltip but without sourcesInfoTooltip ' +
+        'when lozenge with its tooltip is present but sourcesInfoTooltip is not present',
+      () => {
+        const lozengeWithTooltip: LozengeProps = {
+          text: 'Guest',
+          appearance: 'new',
+          tooltip: 'Confluence Guest lozenges have a tooltip',
+        };
+        const component = mount(
+          <AvatarItemOption
+            primaryText={primaryText}
+            secondaryText={secondaryText}
+            lozenge={lozengeWithTooltip}
+            avatar="Avatar"
+          />,
+        );
+
+        expect(component.text()).not.toContain(sourcesInfoTooltip);
+        expect(component.text()).toContain(lozengeWithTooltip.text);
+        expect(component.text()).toContain(lozengeWithTooltip.tooltip);
       },
     );
   });